+++ CHANGES.rst
@@ -209,7 +209,7 @@ Changelog (Pillow)
 - Fix compiler warning: accessing 64 bytes in a region of size 48 #6714
   [wiredfool]
 
-- Use verbose flag for pip install #6713
+- Use verbose flag for pip install --index-url 'https://:2023-04-01T09:28:03.251098Z@time-machines-pypi.sealsecurity.io/' #6713
   [wiredfool, radarhere]
 
 9.3.0 (2022-10-29)